+++ components/NewQuestion.test.js
@@ -4,7 +4,7 @@ import { shallow } from 'enzyme'
 import { Input, Button } from 'reactstrap'
 import NewQuestion from './NewQuestion'
 
-const makeProps = (queueId, userName) => ({
+const makeProps = (queueId, userName = null) => ({
   queueId,
   user: {
     name: userName,
@@ -18,7 +18,7 @@ const doInputChange = (wrapper, name, value) => {
 
 describe('<NewQuestion />', () => {
   test("defaults to empty name if user doesn't have one set", () => {
-    const props = makeProps(1, null)
+    const props = makeProps(1)
     const wrapper = shallow(<NewQuestion {...props} />)
     const nameInput = wrapper.find(Input).find('[name="name"]')
     expect(nameInput.prop('value')).toBe('')
@@ -74,5 +74,21 @@ describe('<NewQuestion />', () => {
     const submitButton = wrapper.find(Button)
     submitButton.simulate('click')
     expect(props.createQuestion).not.toBeCalled()
+  })
+
+  test("shows queue's location for fixed-location queue", () => {
+    const props = makeProps(1)
+    props.queue = { location: 'My Queue Location', fixedLocation: true }
+    const wrapper = shallow(<NewQuestion {...props} />)
+    const locationInput = wrapper.find(Input).find('[name="location"]')
+    expect(locationInput.prop('value')).toBe('My Queue Location')
+  })
+
+  test('disables location field for fixed-location queue', () => {
+    const props = makeProps(1)
+    props.queue = { fixedLocation: true }
+    const wrapper = shallow(<NewQuestion {...props} />)
+    const locationInput = wrapper.find(Input).find('[name="location"]')
+    expect(locationInput.prop('disabled')).toBe(true)
   })
 })